js/xpconnect/tests/chrome/test_bug866823.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=866823
-->
<window title="Mozilla Bug 866823"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=866823"
     target="_blank">Mozilla Bug 866823</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 866823 **/
  SimpleTest.waitForExplicitFinish();
  const Cu = Components.utils;

  function go() {
    var iwin = $('ifr').contentWindow;

    // Make sure that Xray waivers can't be transitively extended onto objects
    // for whom the extender may not waive.
    var sb = new Cu.Sandbox(iwin, { wantXrays: true });
    sb.iwin = iwin;
    ok(Cu.evalInSandbox('iwin.wrappedJSObject.top == iwin.top', sb), "Waiver does not propagate");
    Cu.evalInSandbox('iwin.wrappedJSObject.waiver = iwin.wrappedJSObject', sb);
    ok(iwin.wrappedJSObject.eval('waiver == window'), "Waivers disappear same-compartment");

    // Make sure that COW prototype always happens.
    sb.arrayProto = Array.prototype;
    ok(Cu.evalInSandbox('arrayProto == Array.prototype', sb),
                        "Prototype remapping should happen even when the object isn't used as a prototype");

    // Make sure that prototype remapping happens even for nsEP.
    var sbEP = new Cu.Sandbox(['http://example.org']);
    sb.epArrayProto = sbEP.Array.prototype;
    ok(Cu.evalInSandbox('epArrayProto == Array.prototype', sb),
                        "Prototype remapping should happen for all non-subsuming access");

    SimpleTest.finish();
  }

  ]]>
  </script>
  <iframe id="ifr" onload="go();" type="content" src="http://example.org/tests/js/xpconnect/tests/mochitest/file_empty.html" />
</window>
